All changes requested on my code review has been fixed.

+++ OsiguSDK.Providers/Models/V1/Diagnoses.cs
@@ -9,7 +9,7 @@ namespace OsiguSDK.Providers.Models
         /// Dianosis ID
         /// </summary>
         [JsonProperty(PropertyName = "id")]
-        public decimal Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// Diagnosis
